// front/modulo_cidadao/gts/credenciamento/dto/preenchimento.dto.ts
export interface PreenchimentoForm {
   id?: number;
   nomeProfissional : string;
   cpfProfissional: string;
   telefoneProfissional: string;
   emailProfissional: string;
   formacao: string;
   crmvProfissional: string;
   nomeFantasia: string;
   razaoSocial: string;
   cnpj: string;
   inscricaoEstadual: string;
   telefone: string;
   celular: string;
   email: string;
   crmv: string;
   cepContratante: string;
   logradouroContratante: string;
   numeroContratante: string;
   complementoContratante: string;
   municipioContratante: string;
   ufContratante: string;
}